requirements.txt
crewai
crewai[tools]
fastapi
uvicorn
uuid
pytest-cov
entrypoints==0.3